BaseCollection: create dedicated implementations for Collection
As for https://gitlab.gnome.org/GNOME/vala/issues/548 upcomming
fixes rise errors on non-implemented methods of the same name
in different interfaces.

This fix BaseCollection errors, but now derived classes
don't find how to find the method to override. For example
GomCollection.validate_append() can't be overrided as it is not
found. BaseCollection has now two methods: GomCollection.validate_append()
and Collection.validate_append(), but they are not reachable